@@ -1,7 +1,7 @@
 {
   "name": "multer",
   "description": "Middleware for handling `multipart/form-data`.",
-  "version": "1.4.5-lts.2",
+  "version": "2.0.1",
   "contributors": [
     "Hage Yaapa <captain@hacksparrow.com> (http://www.hacksparrow.com)",
     "Jaret Pfluger <https://github.com/jpfluger>",
@@ -20,25 +20,26 @@
   ],
   "dependencies": {
     "append-field": "^1.0.0",
-    "busboy": "^1.0.0",
-    "concat-stream": "^1.5.2",
-    "mkdirp": "^0.5.4",
+    "busboy": "^1.6.0",
+    "concat-stream": "^2.0.0",
+    "mkdirp": "^0.5.6",
     "object-assign": "^4.1.1",
-    "type-is": "^1.6.4",
-    "xtend": "^4.0.0"
+    "type-is": "^1.6.18",
+    "xtend": "^4.0.2"
   },
   "devDependencies": {
     "deep-equal": "^2.0.3",
-    "express": "^4.13.1",
-    "form-data": "^1.0.0-rc1",
-    "fs-temp": "^1.1.2",
-    "mocha": "^3.5.3",
+    "express": "^4.21.2",
+    "form-data": "^4.0.2",
+    "fs-temp": "^1.2.1",
+    "mocha": "^11.5.0",
+    "nyc": "^15.1.0",
     "rimraf": "^2.4.1",
     "standard": "^14.3.3",
     "testdata-w3c-json-form": "^1.0.0"
   },
   "engines": {
-    "node": ">= 6.0.0"
+    "node": ">= 10.16.0"
   },
   "files": [
     "LICENSE",
@@ -47,6 +48,10 @@
     "lib/"
   ],
   "scripts": {
-    "test": "standard && mocha"
+    "lint": "standard",
+    "lint:fix": "standard --fix",
+    "test": "mocha --reporter spec --exit --check-leaks test/",
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
+    "test-cov": "nyc --reporter=html --reporter=text npm test"
   }
 }
