@@ -1,29 +1,12 @@
 require('dotenv').config();
 const mysql = require('mysql2/promise');
 
-// const koneksi = mysql.createPool({
-//   host: 'localhost',
-//   port: 3306,
-//   user: 'root',
-//   password: 'dedeari123',
-//   database: 'database_film',
-//   waitForConnections: true,
-//   connectionLimit: 10,
-//   queueLimit: 0,
-//   multipleStatements: true
-// });
 
 module.exports = {
-  DB_NAME: 'database_film',
-  DB_USER: 'root',
-  DB_PASSWORD: 'dedeari123',
-  DB_HOST: 'localhost',
-  DB_PORT: 3306,
-  DB_DIALECT: 'mysql'
+  DB_NAME: process.env.DB_NAME,
+  DB_USER: process.env.DB_USER,
+  DB_PASSWORD: process.env.DB_PASSWORD,
+  DB_HOST: process.env.DB_HOST,
+  DB_PORT: process.env.DB_PORT,
+  DB_DIALECT: process.env.DB_DIALECT
 };
-
-// koneksi.getConnection()
-//   .then(() => console.log('✅ Database connected!'))
-//   .catch(err => console.error('❌ Database error:', err.message));
-
-// module.exports = koneksi;
